requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.2
boto3==1.24.88
botocore==1.27.88
cachetools==5.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
django-cors-headers==3.13.0
django-crontab==0.7.1
django-dotenv==1.4.2
django-extensions==3.1.5
django-mailgun-mime==0.1.7
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djantic==0.7.0
dnspython==2.3.0
et-xmlfile==1.1.0
google==3.0.0
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-vision==3.3.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.1
numpy==1.23.5
opencv-python==4.7.0.72
openpyxl==3.0.10
packaging==23.0
pandas==1.5.2
Pillow==9.0.0
proto-plus==1.22.2
protobuf==4.22.0
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pydantic-django==0.1.1
pydot==1.4.2
PyJWT==1.7.1
pymongo==4.3.3
pyparsing==3.0.9
pytesseract==0.3.10
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.4
pyxlsb==1.0.10
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
scraper==0.1.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tabulate==0.9.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
XlsxWriter==3.0.8
